requirements.txt
django==3.2.6
djangorestframework==3.12.4
django_cors_headers==3.2.1
psycopg2-binary==2.9.1
pyyaml==5.4.1
celery[redis]==5.1.2
hiredis==2.0.0
django-filter==2.4.0
coreapi==2.3.3
markdown==3.3.4
django-ckeditor==6.1.0
hashids==1.3.1
django-modeltranslation==0.17.3
django-mptt==0.13.3